requirements.txt
flask==2.3.3
tensorflow==2.13.0
Pillow==10.1.0
numpy==1.24.3
gunicorn==21.2.0
prometheus-client==0.19.0
